+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 
 - **Breaking Change**: cancel code completion when user-defined completer throws an error
 - **Breaking Change**: not allow explicit cast from ``Nothing`` type
+- **Breaking Change**: change operator precedence of ``coproc``
+    - like zsh, ``coproc [pipeline]``
 
 #### Builtin